ai-linkedin-autoposter-starter/requirements.txt
feedparser==6.0.11
PyYAML==6.0.2
beautifulsoup4==4.12.3
requests==2.32.3
sumy==0.11.0
python-dateutil==2.9.0.post0
